many small updates for comments

+++ CRVConditions/inc/CRVROC.hh
@@ -5,20 +5,20 @@
 // a little holder for the triplet of ROC numbering
 //
 
-#include <cstddef>
+#include <cstdint>
 
 namespace mu2e {
 
 class CRVROC {
  public:
-  CRVROC(std::size_t ROC, std::size_t FEB, std::size_t FEBchannel) :
+  CRVROC(std::uint16_t ROC, std::uint16_t FEB, std::uint16_t FEBchannel) :
       _ROC(ROC), _FEB(FEB), _FEBchannel(FEBchannel) {}
-  std::size_t ROC() const { return _ROC; }
-  std::size_t FEB() const { return _FEB; }
-  std::size_t FEBchannel() const { return _FEBchannel; }
-  std::size_t _ROC;
-  std::size_t _FEB;
-  std::size_t _FEBchannel;
+  std::uint16_t ROC() const { return _ROC; }
+  std::uint16_t FEB() const { return _FEB; }
+  std::uint16_t FEBchannel() const { return _FEBchannel; }
+  std::uint16_t _ROC;
+  std::uint16_t _FEB;
+  std::uint16_t _FEBchannel;
 };
 
 }  // namespace mu2e
